replace the login/:token route with reset-password/:token
using login/:token was a possible cause of issues as to reset your
password you needed to first go on the login page to be redirected
to the password reset page.

with this fix, the password reset link no longer makes you redirect
to a page and instead does everything it needs to do on
the password-reset page directly if a reset token is provided.