fix login-required setting not working  (#810)
* fix alloy editor image uploader issues (#757)

"Posting more than one photo is slow, and multiple photos tend
to morph into all showing the same image. More than three photos
 on a idea page and it won't let you save the idea."

Added an if statement to the 'FileSuccess' event inside the alloy
editor that checks if the images 'src' attribute is already set.
if it is, do nothing. this fixes the image morphing and saving
issue.

Issue: #757

* change md-toast z-index to be above the menu badge

* fix login-required setting not working

The setting that lets guest view motions always required guests
to login into the website to view motions even if the setting was
set to false.

This fixes this issue and now guests can now properly view motions
on the website if the setting is turned off.

Issue: #805